erste testbare und weitgehend funktionierende version.
PlayTimer verkuerzt in Main reingeschrieben, PlayTimer.java geloescht.
fps erhoeht. pixels move-methode funktioniert jetzt unabhaengig von den fps, das spiel laeuft also immer gleich schnell.
plattformen sind in einem vector, es werden einige zufaellige plattformen erzeugt.
korrekte (und kommentierte!) kollisionserkennung und reaktion.

+++ trunk/Platform.java
@@ -8,14 +8,14 @@
 public class Platform
 {
     // Instanzvariablen - ersetzen Sie das folgende Beispiel mit Ihren Variablen
-    public int posX;
-    public int posY;
+    public double posX;
+    public double posY;
     public int size;
 
     /**
      * Konstruktor für Objekte der Klasse Platform
      */
-    public Platform(int x, int y, int l)
+    public Platform(double x, double y, int l)
     {
         posX = x;
         posY = y;
